ref: Invert quotas dependency
