+++ relay-dynamic-config/Cargo.toml
@@ -14,12 +14,11 @@ default = []
 processing = []
 
 [dependencies]
-# TODO: Move types used by ProjectConfig from their respective crates to this one. Eventually,
-# this crate should not depend on any sibling crates, except maybe relay-common.
 relay-auth = { path = "../relay-auth" }
 relay-common = { path = "../relay-common" }
 relay-filter = { path = "../relay-filter" }
 relay-general = { path = "../relay-general" }
+relay-quotas = { path = "../relay-quotas" }
 relay-sampling = { path = "../relay-sampling" }
 serde = { version = "1.0.114", features = ["derive"] }
 serde_json = "1.0.55"
