+++ relay-quotas/Cargo.toml
@@ -21,7 +21,6 @@ redis = [
 [dependencies]
 relay-common = { path = "../relay-common" }
 relay-log = { path = "../relay-log", optional = true }
-relay-dynamic-config = { path = "../relay-dynamic-config" }
 relay-redis = { path = "../relay-redis", optional = true }
 serde = { version = "1.0.114", features = ["derive"] }
 smallvec = { version = "1.4.0", features = ["serde"] }
