+++ relay-quotas/src/lib.rs
@@ -11,11 +11,11 @@
 /// typically happens for disabled keys, projects, or organizations.
 const REJECT_ALL_SECS: u64 = 60;
 
+mod quota;
 mod rate_limit;
-mod scoping;
 
+pub use self::quota::*;
 pub use self::rate_limit::*;
-pub use self::scoping::*;
 
 #[cfg(feature = "redis")]
 mod redis;
