+++ relay-quotas/src/redis.rs
@@ -7,9 +7,9 @@ use relay_common::UnixTimestamp;
 use relay_log::protocol::value;
 use relay_redis::{redis::Script, RedisError, RedisPool};
 
+use crate::quota::{ItemScoping, Quota, QuotaScope};
 use crate::rate_limit::{RateLimit, RateLimits, RetryAfter};
-use crate::{ItemScoping, REJECT_ALL_SECS};
-use relay_dynamic_config::quota::{Quota, QuotaScope};
+use crate::REJECT_ALL_SECS;
 
 /// The `grace` period allows accomodating for clock drift in TTL
 /// calculation since the clock on the Redis instance used to store quota
@@ -188,7 +188,7 @@ impl RedisRateLimiter {
         let mut rate_limits = RateLimits::new();
 
         for quota in quotas {
-            if !item_scoping.matches(quota) {
+            if !quota.matches(item_scoping) {
                 // Silently skip all quotas that do not apply to this item.
             } else if quota.limit == Some(0) {
                 // A zero-sized quota is strongest. Do not call into Redis at all, and do not
@@ -260,9 +260,8 @@ mod tests {
     use relay_common::{ProjectId, ProjectKey};
     use relay_redis::{redis::Commands, RedisConfigOptions};
 
+    use crate::quota::{DataCategories, DataCategory, ReasonCode, Scoping};
     use crate::rate_limit::RateLimitScope;
-    use crate::Scoping;
-    use relay_dynamic_config::quota::{DataCategories, DataCategory, ReasonCode};
 
     use super::*;
